chore: Remove separate CRD crate
