--- rust/crd/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "stackable-kafka-crd"
-description = "Contains the Apache Kafka CRD structs and utilities"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-repository.workspace = true
-publish = false
-
-[dependencies]
-indoc.workspace = true
-semver.workspace = true
-serde_json.workspace = true
-serde.workspace = true
-snafu.workspace = true
-stackable-operator.workspace = true
-strum.workspace = true
-tracing.workspace = true
-
-[dev-dependencies]
-serde_yaml.workspace = true
-rstest.workspace = true
