+++ rust/operator-binary/Cargo.toml
@@ -9,7 +9,7 @@ repository.workspace = true
 publish = false
 
 [dependencies]
-stackable-kafka-crd = { path = "../crd" }
+indoc.workspace = true
 anyhow.workspace = true
 clap.workspace = true
 const_format.workspace = true
@@ -25,6 +25,7 @@ tracing.workspace = true
 
 [dev-dependencies]
 serde_yaml.workspace = true
+rstest.workspace = true
 
 [build-dependencies]
 built.workspace = true
