+++ rust/operator-binary/src/crd/affinity.rs
@@ -3,7 +3,7 @@ use stackable_operator::{
     k8s_openapi::api::core::v1::PodAntiAffinity,
 };
 
-use crate::{KafkaRole, APP_NAME};
+use crate::crd::{KafkaRole, APP_NAME};
 
 pub fn get_affinity(cluster_name: &str, role: &KafkaRole) -> StackableAffinityFragment {
     StackableAffinityFragment {
