+++ rust/operator-binary/src/kerberos.rs
@@ -1,8 +1,4 @@
 use snafu::{ResultExt, Snafu};
-use stackable_kafka_crd::{
-    security::KafkaTlsSecurity, KafkaRole, LISTENER_BOOTSTRAP_VOLUME_NAME,
-    LISTENER_BROKER_VOLUME_NAME, STACKABLE_KERBEROS_DIR, STACKABLE_KERBEROS_KRB5_PATH,
-};
 use stackable_operator::builder::{
     self,
     pod::{
@@ -13,6 +9,11 @@ use stackable_operator::builder::{
         },
         PodBuilder,
     },
+};
+
+use crate::crd::{
+    security::KafkaTlsSecurity, KafkaRole, LISTENER_BOOTSTRAP_VOLUME_NAME,
+    LISTENER_BROKER_VOLUME_NAME, STACKABLE_KERBEROS_DIR, STACKABLE_KERBEROS_KRB5_PATH,
 };
 
 #[derive(Snafu, Debug)]
