+++ rust/operator-binary/src/operations/graceful_shutdown.rs
@@ -1,8 +1,9 @@
 use std::collections::BTreeMap;
 
 use snafu::{ResultExt, Snafu};
-use stackable_kafka_crd::KafkaConfig;
 use stackable_operator::builder::pod::PodBuilder;
+
+use crate::crd::KafkaConfig;
 
 #[derive(Debug, Snafu)]
 pub enum Error {
